+++ cmd/release/main.go
@@ -19,7 +19,7 @@ const (
 
 const changelogTemplate = `---
 slug: /changelog/
-sidebar_position: 9
+sidebar_position: 14
 ---`
 
 var (